% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward.r
\name{backward}
\alias{backward}
\title{Backward algorithm}
\usage{
backward(modele)
}
\arguments{
\item{modele}{the forward algorithm results}
}
\value{
This function returns the the probabilities "delta" and "phi" in a matrix.
}
\description{
Backward algorithm
}
