% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.r
\name{EM}
\alias{EM}
\title{Baum-Welch EM algorithm.}
\usage{
EM(
  theta,
  obs,
  modele.fun,
  M.step.fun,
  max.iter = 100,
  trace.theta = TRUE,
  epsilon = 1e-05,
  reltol = sqrt(.Machine$double.eps),
  criteria = c("reltol", "eps")
)
}
\arguments{
\item{theta}{the initial parameters}

\item{obs}{the observations data}

\item{modele.fun}{a model function}

\item{M.step.fun}{a M step function :
- input : obs, backward
- output : theta}

\item{max.iter}{maximum number of iteration of the algorithm (default is 100)}

\item{trace.theta}{whether you want to keep theta estimation for each iteration (default is TRUE)}

\item{epsilon}{if criteria = "eps", algorithm converges if the difference with the previous iteration is lower than eps (default is 1e-5)}

\item{reltol}{if criteria = "reltol", constant related to the stopping criterion, often depending on machine precision (default is sqrt(.Machine$double.eps))}

\item{criteria}{stopping criterion (default is reltol)}
}
\description{
Baum-Welch EM algorithm.
}
\seealso{
\code{\link{forward}}

\code{\link{backward}}
}
