% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.r
\name{forward}
\alias{forward}
\title{Forward algorithm}
\usage{
forward(modele)
}
\arguments{
\item{modele}{a model function}
}
\value{
A list with components `alpha`, `beta`, and `likelihood`. `alpha` and `beta` are matrices with the
the forward probabilities, a and `likelihood` is the log-likelihood.
}
\description{
Forward algorithm
}
