% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QNEM.r
\name{QNEM}
\alias{QNEM}
\title{QNEM algorithm.}
\usage{
QNEM(
  theta,
  obs,
  modele.fun,
  M.step.fun,
  max.iter = 100,
  upper,
  lower,
  trace.theta = TRUE,
  reltol = sqrt(.Machine$double.eps),
  nb.em,
  verbose = FALSE
)
}
\arguments{
\item{theta}{the initial parameters}

\item{obs}{the observations data}

\item{modele.fun}{a model function}

\item{M.step.fun}{a M step function :
- input : obs, backward
- output : theta}

\item{max.iter}{maximum number of iteration of the algorithm (default is 100)}

\item{upper}{the upper bound of the space of parameters}

\item{lower}{the lower bound of the space of parameters}

\item{trace.theta}{whether you want to keep theta estimation for each iteration (default is TRUE)}

\item{reltol}{if criteria = "reltol", constant related to the stopping criterion, often depending on machine precision (default is sqrt(.Machine$double.eps))}

\item{nb.em}{the number of Baum-Welch EM algorithm to perform before switching to BFGS algorithm}

\item{verbose}{if \code{TRUE}, displays information on the process (default is FALSE)}
}
\value{
This function returns the final estimation of the parameters in "theta", the final negative likelihood in "neg.ll", the number of iterations in "iter" and the number of forward and backward algorithms steps in resp. "forward" and "backward". If trace.theta = TRUE, it will also return the parameters estimated for each iteration in "Theta".
}
\description{
QNEM algorithm.
}
\seealso{
\code{\link{modele_derivatives}}

\code{\link{forward_ll}}

\code{\link{backward}}

\code{\link{H_update}}

\code{\link{restrict_inverse}}
}
