% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasi_newton.r
\name{quasi_newton}
\alias{quasi_newton}
\title{Quasi-Newton L-BFGS-B algorithm.}
\usage{
quasi_newton(theta, obs, modele.fun, trace = FALSE, ...)
}
\arguments{
\item{theta}{the initial parameters}

\item{obs}{the observations data}

\item{modele.fun}{a model function}

\item{trace}{whether you want to keep theta estimation for each iteration (default is FALSE)}
}
\value{
This function returns the same values than the optim function. Moreover, the number of forward algorithm steps can be found in "nb.fw".
}
\description{
Quasi-Newton L-BFGS-B algorithm.
}
\details{
You can also add optim function parameters.
}
\seealso{
\code{\link{neg_log_likelihood_gradient}}
}
