% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward.r
\name{forward}
\alias{forward}
\title{Forward algorithm}
\usage{
forward(modele)
}
\arguments{
\item{modele}{a model function}
}
\value{
This function returns the the probabilities "alpha" and "beta" in a matrix and the log-likelihood in "likelihood".
}
\description{
Forward algorithm
}
